Sources/Outcompute.Toolkit/Expressions/IncrementWireExpression.cs
using Outcompute.Toolkit.Expressions.Visitors;

namespace Outcompute.Toolkit.Expressions;

/// <summary>
/// A <see cref="WireExpression"/> that represents the incrementing of the expression value by 1.
/// </summary>
public sealed record class IncrementWireExpression : WireExpression
{
    internal IncrementWireExpression(WireExpression expression)
    {
        Guard.IsNotNull(expression);

        Expression = expression;
    }

    /// <summary>
    /// The expression to convert.
    /// </summary>
    public WireExpression Expression { get; }

    /// <summary>
    /// Makes the specified visitor visit the current expression using the correct overload.
    /// </summary>
    protected internal override WireExpression Accept(WireExpressionVisitor visitor) => visitor.VisitIncrement(this);
}

public partial record class WireExpression
{
    /// <summary>
    /// Creates a <see cref="WireExpression"/> that represents the incrementing of the expression value by 1.
    /// </summary>
    public static IncrementWireExpression Increment(WireExpression expression) => new(expression);
}